requirements.txt
datasets>=2.10.0
anthropic>=0.3.0
numpy
pandas
tqdm
python-dotenv
pytest
